Fix browser auth when the application restarts
The web application was getting into a state where it would reload the various elements from local cache and then try to execute them. But becuase it was no longer authorised, it would end up getting a series of "WWW-Authenticate" headers in responses and would then get into a knot of needing to dismiss them all.

This fix forces a check whenever a route is modified. This check will see if we are authorised already and if not, take us to a login page.